+++ docs/changelog.rst
@@ -5,11 +5,18 @@ Changelog
 versions 0.1.*
 ~~~~~~~~~~~~~~
 
+version 0.0.6_
+^^^^^^^^^^^^^^
+
+- Prepare better tests for console application
+- More options to create arguments for template
+- Arguments you can add from command application
+
 version 0.0.5_
 ^^^^^^^^^^^^^^
 
 - Documentation in Sphinx
-- Fix problem with templates in Pypi package
+- Fix problem with templates in PyPi package
 
 version 0.0.3_
 ^^^^^^^^^^^^^^
@@ -30,6 +37,7 @@ version 0.0.1_
 - Console program to create new program
 
 
+.. _0.0.6: https://github.com/rafyco/templateme/releases/v0.0.6
 .. _0.0.5: https://github.com/rafyco/templateme/releases/v0.0.5
 .. _0.0.3: https://github.com/rafyco/templateme/releases/v0.0.3
 .. _0.0.2: https://github.com/rafyco/templateme/releases/v0.0.2